InstancedStaticMesh Custom Data
+ Exposed per instance custom data to BP

This is a fun one! In UE 4.25 you can assign custom data (as floats) to individual instances of an instanced static mesh. This data is exposed to the material editor, but not to BP. This node lets you get that data, allowing you to define custom logic for individual instances!

This will only work on UE 4.25 or later. If it won't build for previous version anymore, delete the GetCustomDataValue() function from InnerLoopFunctions.cpp and InnerLoopFunctions.h.